Only release the jars which was in the previous release
This prevents the compat jars from being released to studio, where they are no
longer used (ag/9817255).

+++ tools/update_prebuilds_in_android.py
@@ -60,11 +60,13 @@ def copy_targets(root, target_root, srcs, dests, maps=False):
   for i in range(len(srcs)):
     src = os.path.join(root, srcs[i])
     dest = os.path.join(target_root, 'prebuilts', 'r8', dests[i])
-    print 'Copying: ' + src + ' -> ' + dest
-    copyfile(src, dest)
-    if maps:
-      print 'Copying: ' + src + '.map -> ' + dest + '.map'
-      copyfile(src + '.map', dest + '.map')
+    if os.path.exists(dest):
+      print 'Copying: ' + src + ' -> ' + dest
+      copyfile(src, dest)
+      if maps:
+        print 'Copying: ' + src + '.map -> ' + dest + '.map'
+        copyfile(src + '.map', dest + '.map')
+
 
 def copy_jar_targets(root, target_root, jar_targets, maps):
   srcs = map((lambda t: t[0] + '.jar'), jar_targets)
